+++ web/edit/server/iptables/index.php
@@ -1,9 +1,0 @@
-<?php
-$TAB = "SERVER";
-
-// Main include
-include $_SERVER["DOCUMENT_ROOT"] . "/inc/main.php";
-
-// Check user
-header("Location: /list/firewall");
-exit();