+++ web/templates/includes/jump-to-top-link.php
@@ -1,8 +1,0 @@
-<a
-	href="#top"
-	class="button button-secondary button-circle button-floating button-floating-top"
-	title="<?= _("Top") ?>"
->
-	<i class="fas fa-arrow-up"></i>
-	<span class="u-hidden"><?= _("Top") ?></span>
-</a>