+++ web/templates/includes/login-footer.php
@@ -1,4 +1,0 @@
-	</div> <?php // Closes `<div class="app">` in header.php ?>
-
-</body>
-</html>