+++ web/templates/includes/title.php
@@ -1,3 +1,0 @@
-<meta charset="utf-8">
-<title><?= display_title($TAB) ?></title>
-<meta name="viewport" content="width=device-width, initial-scale=1">