+++ web_v2/client/vite.config.js
@@ -7,7 +7,7 @@ export default defineConfig({
     port: 3000,
     proxy: {
       '/api': {
-        target: 'https://localhost:9093',
+        target: 'https://localhost:8083',
         changeOrigin: true,
         secure: false
       }